src/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
celery==5.4.0
redis==5.0.8
httpx==0.27.2
requests==2.32.3
SQLAlchemy[asyncio]==2.0.32
asyncpg==0.29.0
psycopg2-binary==2.9.10
alembic==1.13.2
python-pptx==1.0.2
mammoth==1.8.0
PyMuPDF==1.24.9
pdfminer.six==20231228
prometheus-client==0.20.0

pydantic==2.8.2
structlog==24.1.0

structlog
pydantic-settings==2.4.0
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-instrumentation-asgi>=0.48b0
opentelemetry-instrumentation-requests>=0.48b0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-httpx==0.48b0
opentelemetry-instrumentation-redis==0.48b0
opentelemetry-instrumentation-logging==0.48b0
PyJWT[crypto]==2.8.0
